requirements.txt
fastapi[all]
pre_commit
requests
uvicorn
pydantic
httpx
python-dotenv
click
starlette
pydantic_core
typing_inspection
typing_extensions
sniffio
annotated_types
anyio
h11
cfgv
identify
PyYAML